pyproject.toml
[project]
name = "argolocks"
version = "0.1.0"
description = "ArgoCD deployment approval gates via Slack"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "slack-sdk>=3.33.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
argolocks = "argolocks.main:main"
